can go create posts from frontend
Now it's possible to upload images and fill out forms through the upload process.

PostController can connect uploaded media with post.

CAUTION!!!!
Needs cleaning and error handling in many places. Next thing to implement is redirects and messages, so the user has an idea what's going on. Make sure to reset some of the session data after the creation is complete or failed to avoid conflicts.